Remove utils module
Generally a `utils` module is often somewhat of a code smell, since its
generic name robs its members of more descriptive namespaces.

To avoid this, the only function currently present in the `utils` module
has been moved to the more appropriate `print` module and the module
itself was deleted.
