+++ cli/src/print.rs
@@ -3,6 +3,7 @@ use std::io::{self, Write};
 
 use ansi_term::Color::{Blue, Cyan};
 use clap::Command;
+use prettytable::format;
 use serde::Serialize;
 use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
 
@@ -108,4 +109,16 @@ pub fn truncate(text: &str, max_length: usize) -> Cow<str> {
     } else {
         Cow::Borrowed(text)
     }
+}
+
+pub fn table_format(left_pad: usize, right_pad: usize) -> format::TableFormat {
+    format::FormatBuilder::new()
+        .column_separator(' ')
+        .borders(' ')
+        .separators(
+            &[format::LinePosition::Top, format::LinePosition::Bottom],
+            format::LineSeparator::new(' ', ' ', ' ', ' '),
+        )
+        .padding(left_pad, right_pad)
+        .build()
 }
